[REMEDY-219] Fix remediations loading errors (#327)
To reproduce:
Navigate to `/insights`
Click on Remediations in the side nav (may have to do this multiple times)

This issue is intermittent. If the page loads properly, then you'll need to navigate back to `/insights` and reload the page and start the process all over again. Eventually, you'll get the errors.

You will also see this issue of you navigate to `/insights` and then click the "All apps and services" button in the header and choose "Remediations" from the "Ansible Automation Platform" section.

This PR fixes that issue.

Co-authored-by: Michael Johnson <micjohns@newbeginnings.usersys.redhat.com>