fix(RHINENG-4945): Fix bulk select on archive
To repro:
Navigate to Remediations
Ensure only non-archived playbooks are being shown via toolbar action kebab
Select checkbox of a playbook
Archive playbook from card action kebab

The playbook remains selected (see Bulk Selector) and I am able to delete it even though it is not shown in the table

With this PR, when only non-archived playbooks are being shown, and we archive a selected playbook, that selection should not remain.

On the other hand, if we are viewing all playbooks (archived and non-archived) and we archive a selected playbook, the selection should remain.