fix(RHINENG-8088): refactor select playbook typeahead (#422)
* fix(RHINENG-8088): refactor select playbook typeahead

To repro:
Choose and click a system in inventory
From the system details page, click the Vulnerability tab
Pick CVE(s) and click on Remediate button
When modal opens - click in the "Select playbook" box

This box should allow typeahead, which filters the existing playbooks as you type. But this feature seems to have been removed from the PF component we were using and have been moved to the Select component.

This PR refactors this box to allow typeahead.

* Add suggested fixes and more tests

---------

Co-authored-by: Michael Johnson <micjohns@micjohns-thinkpadp16vgen1.rmtusnc.csb>